requirements.txt
numpy
opencv-python
tensorflow-gpu==1.13.1
keras=2.2.4
